Nodered Coding
This all code is in JSON  format. It use to access data through MQTT  and passes it through JSON to convert it. Then it stores data and do somer analysis and processing part. If it senses  some crititcal situation then send command to turn ON siren and trigger trilio API.